LangChain/Projects/Youtube_Video_Summarizer/requirements.txt
streamlit
langchain==0.2.16
langchain-openai
openai
youtube-transcript-api==0.6.2
python-dotenv
pytube
